+++ src/main/java/nextstep/subway/path/domain/policy/AgePaymentPolicy.java
@@ -1,0 +1,17 @@
+package nextstep.subway.path.domain.policy;
+
+import nextstep.subway.path.domain.Cost;
+import nextstep.subway.path.domain.CostByAge;
+import nextstep.subway.path.dto.CostRequest;
+
+public class AgePaymentPolicy implements PaymentPolicy {
+    private PaymentPolicy next;
+
+    @Override
+    public CostRequest cost(CostRequest costRequest) {
+        Cost cost = CostByAge.applyDiscount(costRequest.getCost(), costRequest.getAge());
+        costRequest.updateCost(cost);
+
+        return costRequest;
+    }
+}
